+++ tests/Justimmo/Tests/TestCase.php
@@ -4,7 +4,7 @@ namespace Justimmo\Tests;
 
 use PHPUnit\Framework\TestCase as PHPUnitTestCase;
 
-class TestCase extends PHPUnitTestCase
+abstract class TestCase extends PHPUnitTestCase
 {
     protected function getFixtures($file)
     {